// src/app/(authenticated)/captura/types/types.ts
export interface Processo {
  id: number;
  descricaoOrgaoJulgador: string;
  classeJudicial: string;
  numero: number;
  numeroProcesso: string;
  segredoDeJustica: boolean;
  codigoStatusProcesso: string;
  prioridadeProcessual: number;
  nomeParteAutora: string;
  qtdeParteAutora: number;
  nomeParteRe: string;
  qtdeParteRe: number;
  dataAutuacao: string;
  juizoDigital: boolean;
  dataArquivamento?: string;
  dataProximaAudiencia?: string | null;
  temAssociacao: boolean;
}